styling Admin page with edit, delte and add
