+++ src/Components/Admin/AdminBoard.module.css
@@ -1,0 +1,110 @@
+.board {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: .5rem;
+    height: 100vh;
+    padding: 3rem;
+}
+
+.board-header {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    font-size: 1.6rem;
+}
+
+.add-btn {
+    border-radius: 100px;
+    padding: .65rem .65rem .5rem .65rem;
+    background-color: #936EFE;
+    color: #fff;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+}
+
+.add-btn:hover {
+    transform: scale(106%);
+    cursor: pointer;
+    transition: 150ms;
+}
+
+.add-btn:active {
+  transform: translate(0, .05rem);
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.30);
+}
+
+.board-title {
+    width: 100%;
+    max-width: 1250px;
+    padding: 0 1rem;
+    font-size: 2rem;
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+}
+
+.board-container {
+    display: flex;
+    align-items: flex-start;
+    justify-content: space-between;
+    background-color: white;
+    width: 100%;
+    height: 70%;
+    border-radius: 12px;
+    max-width: 1250px;
+    border-radius: 10px;
+    gap: 3rem;
+    box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
+}
+
+.board-data-container {
+    display: flex;
+    flex-direction: column;
+    gap: 2.5rem;
+    width: 100%;
+    height: 100%;
+    padding: 2.5rem;
+    overflow: auto;
+    -ms-overflow-style: none;  /* IE and Edge */
+    scrollbar-width: none;  /* Firefox */
+}
+
+.product-container {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.product-image {
+    width: 16rem;
+    height: 10rem;
+    object-fit: cover;
+    background-repeat: no-repeat;
+    object-position: center 70%;
+    border-radius: 8px;
+}
+
+.product-btns {
+    margin-left: auto;
+}
+
+.delete-btn,
+.edit-btn {
+    color: #333;
+    margin-left: 1rem;
+    background-color: transparent;
+}
+
+.delete-btn:hover,
+.edit-btn:hover {
+    cursor: pointer;
+  }
+
+button {
+    border: none;
+    color: #fafafa;
+    font-size: 1.2rem;
+}
