+++ src/Components/UI/Modal.module.css
@@ -10,7 +10,7 @@
   
   .modal {
     position: fixed;
-    top: 4vh;
+    top: 10%;
     left: 5%;
     width: 90%;
     background-color: white;
